factor out image glyph base class
