+++ bokehjs/src/lib/models/glyphs/image.ts
@@ -1,4 +1,4 @@
-import {XYGlyph, XYGlyphView, XYGlyphData} from "./xy_glyph"
+import {ImageBase, ImageBaseView, ImageDataBase} from "./image_base"
 import {ColorMapper} from "../mappers/color_mapper"
 import {LinearColorMapper} from "../mappers/linear_color_mapper"
 import {Class} from "core/class"
@@ -6,26 +6,12 @@ import {Arrayable} from "core/types"
 import * as p from "core/properties"
 import {concat} from "core/util/array"
 import {Context2d} from "core/util/canvas"
-import {SpatialIndex} from "core/util/spatial"
 import * as hittest from "core/hittest"
 import {Selection} from "../selections/selection"
 import {PointGeometry} from "core/geometry"
+import { ImageRGBAData } from './image_rgba'
 
-// XXX: because ImageData is a global
-export interface _ImageData extends XYGlyphData {
-  image_data: Arrayable<HTMLCanvasElement>
-
-  _image: Arrayable<Arrayable<number> | number[][]>
-  _dw: Arrayable<number>
-  _dh: Arrayable<number>
-
-  _image_shape?: Arrayable<[number, number]>
-
-  sw: Arrayable<number>
-  sh: Arrayable<number>
-}
-
-export interface ImageView extends _ImageData {}
+export interface ImageView extends ImageDataBase {}
 
 export interface ImageIndex {
   index: number
@@ -34,7 +20,7 @@ export interface ImageIndex {
   flat_index: number
 }
 
-export class ImageView extends XYGlyphView {
+export class ImageView extends ImageBaseView {
   model: Image
   visuals: Image.Visuals
 
@@ -53,32 +39,6 @@ export class ImageView extends XYGlyphView {
       this._set_data()
       this.renderer.plot_view.request_render()
     }
-  }
-
-  _index_data(): SpatialIndex {
-    const points = []
-    for (let i = 0, end = this._x.length; i < end; i++) {
-      const [l, r, t, b] = this._lrtb(i)
-      if (isNaN(l + r + t + b) || !isFinite(l + r + t + b)) {
-        continue
-      }
-      points.push({minX: l, minY: b, maxX: r, maxY: t, i})
-    }
-    return new SpatialIndex(points)
-  }
-
-  _lrtb(i: number) : [number, number, number, number]{
-    const xr = this.renderer.xscale.source_range
-    const x1 = this._x[i]
-    const x2 = xr.is_reversed ? x1 - this._dw[i] : x1 + this._dw[i]
-
-    const yr = this.renderer.yscale.source_range
-    const y1 = this._y[i]
-    const y2 = yr.is_reversed ? y1 - this._dh[i] : y1 + this._dh[i]
-
-    const [l,r] = x1 < x2 ? [x1,x2] : [x2,x1]
-    const [b,t] = y1 < y2 ? [y1,y2] : [y2,y1]
-    return [l, r, t, b]
   }
 
   _image_index(index : number, x: number, y : number) : ImageIndex {
@@ -110,14 +70,7 @@ export class ImageView extends XYGlyphView {
   }
 
   protected _set_data(): void {
-    if (this.image_data == null || this.image_data.length != this._image.length)
-      this.image_data = new Array(this._image.length)
-
-    if (this._width == null || this._width.length != this._image.length)
-      this._width = new Array(this._image.length)
-
-    if (this._height == null || this._height.length != this._image.length)
-      this._height = new Array(this._image.length)
+    this._set_width_heigh_data()
 
     const cmap = this.model.color_mapper.rgba_mapper
 
@@ -135,51 +88,13 @@ export class ImageView extends XYGlyphView {
         this._width[i] = _image[0].length
       }
 
-      const _image_data = this.image_data[i]
-      let canvas: HTMLCanvasElement
-      if (_image_data != null && _image_data.width == this._width[i] &&
-                                 _image_data.height == this._height[i])
-        canvas = _image_data
-      else {
-        canvas = document.createElement('canvas')
-        canvas.width = this._width[i]
-        canvas.height = this._height[i]
-      }
-
-      const ctx = canvas.getContext('2d')!
-      const image_data = ctx.getImageData(0, 0, this._width[i], this._height[i])
       const buf8 = cmap.v_compute(img)
-      image_data.data.set(buf8)
-      ctx.putImageData(image_data, 0, 0)
-      this.image_data[i] = canvas
+      this._set_image_data_from_buffer(i, buf8)
+
     }
   }
 
-  protected _map_data(): void {
-    switch (this.model.properties.dw.units) {
-      case "data": {
-        this.sw = this.sdist(this.renderer.xscale, this._x, this._dw, 'edge', this.model.dilate)
-        break
-      }
-      case "screen": {
-        this.sw = this._dw
-        break
-      }
-    }
-
-    switch (this.model.properties.dh.units) {
-      case "data": {
-        this.sh = this.sdist(this.renderer.yscale, this._y, this._dh, 'edge', this.model.dilate)
-        break
-      }
-      case "screen": {
-        this.sh = this._dh
-        break
-      }
-    }
-  }
-
-  protected _render(ctx: Context2d, indices: number[], {image_data, sx, sy, sw, sh}: _ImageData): void {
+  protected _render(ctx: Context2d, indices: number[], {image_data, sx, sy, sw, sh}: ImageRGBAData): void {
     const old_smoothing = ctx.getImageSmoothingEnabled()
     ctx.setImageSmoothingEnabled(false)
 
@@ -213,7 +128,7 @@ const Greys9 = () => ["#000000", "#252525", "#525252", "#737373", "#969696", "#b
 export namespace Image {
   export type Attrs = p.AttrsOf<Props>
 
-  export type Props = XYGlyph.Props & {
+  export type Props = ImageBase.Props & {
     image: p.NumberSpec
     dw: p.DistanceSpec
     dh: p.DistanceSpec
@@ -222,12 +137,12 @@ export namespace Image {
     color_mapper: p.Property<ColorMapper>
   }
 
-  export type Visuals = XYGlyph.Visuals
+  export type Visuals = ImageBase.Visuals
 }
 
 export interface Image extends Image.Attrs {}
 
-export class Image extends XYGlyph {
+export class Image extends ImageBase {
   properties: Image.Props
   default_view: Class<ImageView>
 
@@ -240,11 +155,6 @@ export class Image extends XYGlyph {
     this.prototype.default_view = ImageView
 
     this.define<Image.Props>({
-      image:        [ p.NumberSpec       ], // TODO (bev) array spec?
-      dw:           [ p.DistanceSpec     ],
-      dh:           [ p.DistanceSpec     ],
-      dilate:       [ p.Boolean,   false ],
-      global_alpha: [ p.Number,    1.0   ],
       color_mapper: [ p.Instance,  () => new LinearColorMapper({palette: Greys9()}) ],
     })
   }
