+++ bokehjs/src/lib/models/glyphs/image_rgba.ts
@@ -1,26 +1,15 @@
-import {XYGlyph, XYGlyphView, XYGlyphData} from "./xy_glyph"
+import {ImageBase, ImageBaseView, ImageDataBase} from "./image_base"
 import {Arrayable, TypedArray} from "core/types"
 import {Class} from "core/class"
 import * as p from "core/properties"
 import {concat} from "core/util/array"
 import {Context2d} from "core/util/canvas"
 
-export interface ImageRGBAData extends XYGlyphData {
-  image_data: Arrayable<HTMLCanvasElement>
-
-  _image: Arrayable<TypedArray | number[][]>
-  _dw: Arrayable<number>
-  _dh: Arrayable<number>
-
-  _image_shape?: Arrayable<[number, number]>
-
-  sw: Arrayable<number>
-  sh: Arrayable<number>
-}
+export interface ImageRGBAData extends ImageDataBase {}
 
 export interface ImageRGBAView extends ImageRGBAData {}
 
-export class ImageRGBAView extends XYGlyphView {
+export class ImageRGBAView extends ImageBaseView {
   model: ImageRGBA
   visuals: ImageRGBA.Visuals
 
@@ -33,14 +22,7 @@ export class ImageRGBAView extends XYGlyphView {
   }
 
   protected _set_data(indices: number[] | null): void {
-    if (this.image_data == null || this.image_data.length != this._image.length)
-      this.image_data = new Array(this._image.length)
-
-    if (this._width == null || this._width.length != this._image.length)
-      this._width = new Array(this._image.length)
-
-    if (this._height == null || this._height.length != this._image.length)
-      this._height = new Array(this._image.length)
+    this._set_width_heigh_data()
 
     for (let i = 0, end = this._image.length; i < end; i++) {
       if (indices != null && indices.indexOf(i) < 0)
@@ -64,47 +46,9 @@ export class ImageRGBAView extends XYGlyphView {
         this._width[i] = _image[0].length
       }
 
-      const _image_data = this.image_data[i]
-      let canvas: HTMLCanvasElement
-      if (_image_data != null && _image_data.width == this._width[i] &&
-                                 _image_data.height == this._height[i])
-        canvas = _image_data
-      else {
-        canvas = document.createElement('canvas')
-        canvas.width = this._width[i]
-        canvas.height = this._height[i]
-      }
-
-      const ctx = canvas.getContext('2d')!
-      const image_data = ctx.getImageData(0, 0, this._width[i], this._height[i])
       const buf8 = new Uint8Array(buf)
-      image_data.data.set(buf8)
-      ctx.putImageData(image_data, 0, 0)
-      this.image_data[i] = canvas
-    }
-  }
+      this._set_image_data_from_buffer(i, buf8)
 
-  protected _map_data(): void {
-    switch (this.model.properties.dw.units) {
-      case "data": {
-        this.sw = this.sdist(this.renderer.xscale, this._x, this._dw, "edge", this.model.dilate)
-        break
-      }
-      case "screen": {
-        this.sw = this._dw
-        break
-      }
-    }
-
-    switch (this.model.properties.dh.units) {
-      case "data": {
-        this.sh = this.sdist(this.renderer.yscale, this._y, this._dh, "edge", this.model.dilate)
-        break
-      }
-      case "screen": {
-        this.sh = this._dh
-        break
-      }
     }
   }
 
@@ -136,7 +80,7 @@ export class ImageRGBAView extends XYGlyphView {
 export namespace ImageRGBA {
   export type Attrs = p.AttrsOf<Props>
 
-  export type Props = XYGlyph.Props & {
+  export type Props = ImageBase.Props & {
     image: p.NumberSpec
     dw: p.DistanceSpec
     dh: p.DistanceSpec
@@ -144,12 +88,12 @@ export namespace ImageRGBA {
     dilate: p.Property<boolean>
   }
 
-  export type Visuals = XYGlyph.Visuals
+  export type Visuals = ImageBase.Visuals
 }
 
 export interface ImageRGBA extends ImageRGBA.Attrs {}
 
-export class ImageRGBA extends XYGlyph {
+export class ImageRGBA extends ImageBase {
   properties: ImageRGBA.Props
   default_view: Class<ImageRGBAView>
 
@@ -160,14 +104,6 @@ export class ImageRGBA extends XYGlyph {
   static initClass(): void {
     this.prototype.type = 'ImageRGBA'
     this.prototype.default_view = ImageRGBAView
-
-    this.define<ImageRGBA.Props>({
-      image:        [ p.NumberSpec         ], // TODO (bev) array spec?
-      dw:           [ p.DistanceSpec       ],
-      dh:           [ p.DistanceSpec       ],
-      global_alpha: [ p.Number,      1.0   ],
-      dilate:       [ p.Boolean,     false ],
-    })
   }
 }
 ImageRGBA.initClass()
